IDs are now Longs instead of Ints for MySQL reasons

+++ src/main/kotlin/com/doubleu/kotlintrader/database/Entity.kt
@@ -18,7 +18,7 @@ abstract class Entity {
         registerId(this.id)
     }
 
-    abstract val id: Int
+    abstract val id: Long
 
     val delegateMap = mutableMapOf<KProperty<*>, DatabaseDelegate<*>>()
 
@@ -30,7 +30,7 @@ abstract class Entity {
 
     inline fun <reified T : Entity> reference(
             key: KProperty<T>,
-            property: KProperty<Int>)
+            property: KProperty<Long>)
             : ReferenceDelegate<T> {
         val delegate = ReferenceDelegate(T::class, property)
         delegateMap[key] = delegate
@@ -39,7 +39,7 @@ abstract class Entity {
 
     inline fun <reified T : Entity> mutableReference(
             key: KMutableProperty0<T?>,
-            property: KMutableProperty0<Int?>)
+            property: KMutableProperty0<Long?>)
             : MutableReferenceDelegate<T> {
         val delegate = MutableReferenceDelegate(T::class, property)
         delegateMap[key] = delegate
@@ -50,7 +50,7 @@ abstract class Entity {
     fun <V> property(property: KProperty<V>): Property<V> = (delegateMap[property]?.property
             ?: throw RuntimeException("${property.name} wasn't yet delegated!")) as Property<V>
 
-    protected fun registerId(id: Int) {
+    protected fun registerId(id: Long) {
         // TODO: Use INSERT if not present, think of default values
     }
 }
