+++ src/main/kotlin/com/doubleu/kotlintrader/model/Ereignis.kt
@@ -2,7 +2,7 @@ package com.doubleu.kotlintrader.model
 
 import com.doubleu.kotlintrader.database.Entity
 
-class Ereignis(override val id: Int) : Entity() {
+class Ereignis(override val id: Long) : Entity() {
 
     var beschreibung: String? by delegate(this::beschreibung)
     val beschreibungProperty = property(this::beschreibung)
